Fix for meteor android
Thank you @EddyVerbruggen for your package, I’m using it in a Meteor.js
package and while doing so I stumbled on an issue.

The intent gets cleared even if theres no handler, this pr adds a clear
intent command and runs this when calling the handle.

The issue in Meteor is that theres a update mech (the initial page)
when done it redirects to the actual app - the intent would be cleared
in the initial page and the handle in the app would not trigger.

Let me know if you have additional questions about the fix