Adding Extras bundle to the URL as an json-string
Hi. 
Note: As i cannot run your latest version, it seems to be some corruption in the plugin.xml, i have not verified this on the latest src. only on 4.1.5 and patched the code by hand. But there should at most only be simple compile issues.

Reson for patch: 
I need to catch Extras data from a calling Intent. Hence this patch includes them as a JSON formatted string.
Patch only covers this scenario from a warm start start ( i.e. app is already started) as it't the only time i need it caught.
But it might be extended to the other cases in the same manner.